pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.sdist]
only-include = ["ai_nexus_backend"]
exclude = [
  "/.github",
  "htmlcov",
  ".coverage",
  "coverage.xml",
]
[tool.hatch.build.targets.wheel]
packages = ["./ai_nexus_backend"]
[project]
name = "ai_nexus_backend"
version = "0.1.1"
requires-python = ">=3.9,<=3.13"
description = "Modular component for AI-enabled application search backend."
readme = "README.md"
authors = [
    { name="Rich Leyshon", email="richard.leyshon@digital.justice.gov.uk" },
    { name="Chloe Pugh", email="chloe.pugh@digital.justice.gov.uk" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: MacOS",
]
dependencies = [
    "beautifulsoup4==4.12.3",
    "fastparquet==2024.5.0",
    "pandas==2.2.3",
    "pyarrow==17.0.0",
    "pyprojroot==0.3.0",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    ]

[project.optional-dependencies]
# Add your optional dependencies here
dev = [
    "pytest==8.3.3",
    "hatchling==1.25.0",
    "mockito==1.5.1",
    "pre-commit==4.0.0",
]
[project.urls]
Repository = "https://github.com/ministryofjustice/..."
"Bug tracker" = "https://github.com/ministryofjustice/.../issues"
# `coverage` configurations
[tool.coverage.run]
source = [
    "./ai_nexus_backend"
]
omit = ["**/__init__.py"]
[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:"
]
# `pytest` configurations
[tool.pytest.ini_options]
addopts = [
    "-vv",
    "--doctest-modules"
]
doctest_optionflags = "NORMALIZE_WHITESPACE"
testpaths = [
    "./tests"
]
